pyproject.toml
[tool.poetry]
name = "iot-analytics"
version = "0.1.0"
description = ""
authors = ["Gabriel Chiong <gabrielzhenchiong@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
gunicorn = "^23.0.0"
redis = "^5.2.1"
psycopg2-binary = "^2.9.10"
django-redis = "^5.4.0"
channels = "^4.2.0"
channels-redis = "^4.2.1"
uvicorn = "^0.34.0"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
